make extra InterfaceExporter.export(blob: )
