+++ Sources/ApodiniExtension/ErrorHandling.swift
@@ -1,5 +1,5 @@
 //
-//  ErrorEncoding.swift
+//  ErrorHandling.swift
 //  
 //
 //  Created by Max Obermeier on 25.06.21.
@@ -8,6 +8,26 @@
 import Foundation
 import Apodini
 import ApodiniUtils
+import OpenCombine
+import NIO
+
+public protocol ErrorHandler {
+    associatedtype Output
+    associatedtype Failure
+    
+    func handle(_ error: ApodiniError) -> ErrorHandlingStrategy<Output, Failure>
+}
+
+public enum ErrorHandlingStrategy<Output, Failure> {
+    case graceful(Output)
+    case ignore
+    case abort(Failure)
+}
+
+//public extension EventLoopFuture {
+//    
+//    
+//}
 
 public extension Error {
     var standardMessage: String {
