+++ Sources/ApodiniOpenAPI/OpenAPIInterfaceExporter.swift
@@ -71,6 +71,10 @@ final class OpenAPIInterfaceExporter: InterfaceExporter {
         }
     }
     
+    func export<H>(blob endpoint: Endpoint<H>) where H : Handler, H.Response.Content == Blob {
+        export(endpoint)
+    }
+    
     func finishedExporting(_ webService: WebServiceModel) {
         serveSpecification()
         updateStorage()
